requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-binance==1.0.19
jinja2==3.1.2
httpx==0.24.1
requests==2.31.0
numpy>=1.21.0
python-dotenv==1.0.0
itsdangerous==2.1.2
python-multipart==0.0.6